make sub route & react-query work together
