+++ components/TeacherPortal/Tags.tsx
@@ -9,7 +9,10 @@ import { Button } from '../ui/button';
 import NewTagButton from './Tags/NewTagButton';
 import Tag from './Tags/Tag';
 
-const Tags = ({ tags }: { tags: TagType[] }) => {
+const Tags = () => {
+  const { fetchQuery } = useTeacherTags();
+  const tags: TagType[] = fetchQuery.data || [];
+  console.log('inter::::', fetchQuery.data?.length);
   const { t } = useTranslation('model');
   const { removeTeacherTags, addTeacherTag } = useTeacherTags();
   const { mutateAsync: removeTags } = removeTeacherTags;
