--- components/TeacherPortal/teacher-portal.state.tsx
@@ -1,8 +1,0 @@
-export const portalState = {
-  showing: 'one-time-code' as
-    | 'one-time-code'
-    | 'shared-message'
-    | 'tags'
-    | 'teacher-prompt'
-    | 'settings',
-};
