+++ pages/teacher-portal/index.tsx
@@ -1,141 +1,18 @@
-import { Dispatch, useCallback, useEffect, useRef, useState } from 'react';
-import LoadingBar, { LoadingBarRef } from 'react-top-loading-bar';
+import { GetServerSideProps } from 'next';
 
-import useTeacherTags from '@/hooks/teacherPortal/useTeacherTags';
-
-import { withCommonServerSideProps } from '@/utils/withCommonServerSideProps';
-
-import { Tag } from '@/types/tags';
-
-import Spinner from '@/components/Spinner';
-import OneTimeCodeGeneration from '@/components/TeacherPortal/OneTimeCodeGeneration';
-import SharedMessages from '@/components/TeacherPortal/SharedMessages';
-import Sidebar from '@/components/TeacherPortal/Sidebar';
-import Tags from '@/components/TeacherPortal/Tags';
-import TeacherPrompt from '@/components/TeacherPortal/TeacherPrompt';
-import TeacherSettings from '@/components/TeacherPortal/TeacherSettings';
-import {
-  ShowingChangeAction,
-  TeacherPortalContext,
-} from '@/components/TeacherPortal/teacher-portal.context';
-import { portalState } from '@/components/TeacherPortal/teacher-portal.state';
-import DefaultLayout from '@/components/layout/default';
-
-import { createServerSupabaseClient } from '@supabase/auth-helpers-nextjs';
-
-const TeacherPortal = () => {
-  const [showing, setShowing] = useState(portalState.showing);
-
-  const loadingRef = useRef<LoadingBarRef>(null);
-  const startLoading = useCallback(() => {
-    loadingRef.current?.continuousStart();
-  }, []);
-  const completeLoading = useCallback(() => {
-    loadingRef.current?.complete();
-  }, []);
-
-  const dispatch: Dispatch<ShowingChangeAction> = (action) => {
-    if (action.field === 'showing') {
-      setShowing(action.value);
-    } else {
-      setShowing(portalState.showing);
-    }
-  };
-  const [hasMounted, setHasMounted] = useState(false);
-  useEffect(() => {
-    setHasMounted(true);
-  }, []);
-  return (
-    <DefaultLayout>
-      {hasMounted && (
-        <TeacherPortalContext.Provider
-          value={{
-            state: { showing },
-            dispatch,
-            startLoading,
-            completeLoading,
-          }}
-        >
-          <LoadingBar color={'white'} ref={loadingRef} />;
-          <TeacherPortalContent showing={showing} />
-        </TeacherPortalContext.Provider>
-      )}
-    </DefaultLayout>
-  );
-};
-
-export default TeacherPortal;
-
-interface TeacherPortalContentProps {
-  showing: string;
-}
-
-const TeacherPortalContent: React.FC<TeacherPortalContentProps> = ({
-  showing,
-}) => {
-  const { fetchQuery } = useTeacherTags();
-  const { data, isLoading } = fetchQuery;
-  const tags: Tag[] = data || [];
-
-  return (
-    <div className="fixed inset-0 overflow-y-auto">
-      <div className="flex min-h-full items-center justify-center text-center mobile:block">
-        <div className="bg-neutral-900 w-full tablet:max-w-[90vw] transform overflow-hidden text-left align-middle shadow-xl transition-all text-neutral-200 flex h-[100dvh] tablet:max-h-[unset] !max-w-[unset] !rounded-none">
-          <Sidebar className="bg-neutral-800 flex-shrink-0 flex-grow-0" />
-
-          {isLoading ? (
-            <div className="flex-grow relative flex items-center justify-center">
-              <Spinner size="16px" />
-            </div>
-          ) : (
-            <div className="p-6 flex-grow relative overflow-y-auto">
-              {showing === 'one-time-code' && (
-                <OneTimeCodeGeneration tags={tags} />
-              )}
-              {showing === 'shared-message' && <SharedMessages tags={tags} />}
-              {showing === 'tags' && <Tags tags={tags} />}
-              {showing === 'teacher-prompt' && <TeacherPrompt />}
-              {showing === 'settings' && <TeacherSettings />}
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export const getServerSideProps = withCommonServerSideProps(async (context) => {
-  const supabase = createServerSupabaseClient(context);
-  const {
-    data: { user },
-    error,
-  } = await supabase.auth.getUser();
-
-  if (error || !user) {
-    return {
-      redirect: {
-        destination: '/',
-        permanent: false,
-      },
-    };
-  }
-
-  const { data: profile, error: profileError } = await supabase
-    .from('profiles')
-    .select('*')
-    .eq('id', user?.id)
-    .single();
-
-  if (profileError || !profile || !profile.is_teacher_account) {
-    return {
-      redirect: {
-        destination: '/',
-        permanent: false,
-      },
-    };
-  }
-
+export const getServerSideProps: GetServerSideProps = async (context) => {
   return {
-    props: {},
+    redirect: {
+      destination: '/teacher-portal/one-time-code', // Target route
+      permanent: false, // Temporary redirect, set to true if it's a permanent redirect
+    },
   };
-});
+};
+
+const TeacherPortalIndex = () => {
+  // This component will not actually render because of the server-side redirect,
+  // but you can return a simple component or loading state if needed.
+  return null;
+};
+
+export default TeacherPortalIndex;
