+++ types/teacher-portal-sub-route.ts
@@ -1,0 +1,2 @@
+export type RouteType = string &
+  ('one-time-code' | 'shared-message' | 'tags' | 'teacher-prompt' | 'settings');
